Funcao de adicionar no carrinho dentro da classe de Product
