+++ src/Server/cmdProcess.java
@@ -73,18 +73,4 @@ public class cmdProcess {
 		}
 		return line;
 	}
-
-	public static String CalendarToString (GregorianCalendar date) {
-		return date.get(Calendar.DAY_OF_MONTH) + "/"
-				+(date.get(Calendar.MONTH)+1) + "/"
-				+date.get(Calendar.YEAR);
-	}
-
-	public static GregorianCalendar StringToCalendar (String date) {
-		String[] split_date = date.split("/");
-
-		return new GregorianCalendar(Integer.parseInt(split_date[2]),
-				Integer.parseInt(split_date[1])-1,
-				Integer.parseInt(split_date[0]));
-	}
 }
